Move ExternalTab to gui.preftabs
